Revert to avoid hard to read diff: "Rename AutoscalerFlinkMetrics to AutoscalerMetrics to avoid confusion"
