+++ flink-kubernetes-operator-autoscaler/src/main/java/org/apache/flink/kubernetes/operator/autoscaler/AutoscalerFlinkMetrics.java
@@ -36,9 +36,9 @@ import static org.apache.flink.kubernetes.operator.autoscaler.metrics.ScalingMet
 import static org.apache.flink.kubernetes.operator.autoscaler.metrics.ScalingMetric.RECOMMENDED_PARALLELISM;
 
 /** Autoscaler metrics for observability. */
-public class AutoscalerMetrics {
+public class AutoscalerFlinkMetrics {
 
-    private static final Logger LOG = LoggerFactory.getLogger(AutoscalerMetrics.class);
+    private static final Logger LOG = LoggerFactory.getLogger(AutoscalerFlinkMetrics.class);
     @VisibleForTesting static final String CURRENT = "Current";
     @VisibleForTesting static final String AVERAGE = "Average";
     @VisibleForTesting static final String JOB_VERTEX_ID = "jobVertexID";
@@ -53,7 +53,7 @@ public class AutoscalerMetrics {
 
     private boolean scalingMetricsInitialized;
 
-    public AutoscalerMetrics(MetricGroup metricGroup) {
+    public AutoscalerFlinkMetrics(MetricGroup metricGroup) {
         this.numScalings = metricGroup.counter("scalings");
         this.numErrors = metricGroup.counter("errors");
         this.numBalanced = metricGroup.counter("balanced");
